refactor(url-params): moved the UrlSearchParamsContext outside of the components directory
